pyproject.toml
[project]
name = "firstwork-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "numpy>=2.3.1",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "pydantic[email]>=2.11.7",
    "pymupdf>=1.26.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "typer>=0.16.0",
]
